requirements.txt
opencv-python
pillow
black
pylint
